% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ts.R
\name{mod_ts_ui}
\alias{mod_ts_ui}
\alias{mod_ts_server}
\title{mod_ts_ui and mod_ts_server}
\usage{
mod_ts_ui(id)

mod_ts_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
